+++ src/messages/lib/CommandTimeRun.js
@@ -1,0 +1,10 @@
+const AbstractMessage = require('./AbstractMessage')
+
+class CommandTimeRun extends AbstractMessage {
+	constructor () {
+		super()
+		this.topic = 'command_time_run'
+	}
+}
+
+module.exports = CommandTimeRun