+++ test/core/withAsset.spec.js
@@ -1,7 +1,7 @@
 const { Testkit } = require('../../main')
 const { Network } = Testkit()
 
-describe('Check NetworkInitializer with function (everything together)', function () {
+describe('Check NetworkInitializer with asset function', function () {
 	this.timeout(60000)
 
 	before(async () => {